% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convolutional_model.R
\name{conv_build_model}
\alias{conv_build_model}
\title{Call this function as an addin to insert the example of "Convolutional Model building" at the cursor position.}
\usage{
conv_build_model()
}
\description{
Call this function as an addin to insert the example of "Convolutional Model building" at the cursor position.
}
