% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperModel_class.R
\name{HyperModel}
\alias{HyperModel}
\title{Call this function as an addin to insert "HyperModel class" at the cursor position.}
\usage{
HyperModel()
}
\description{
Call this function as an addin to insert "HyperModel class" at the cursor position.
}
