% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperclass.R
\name{HyperModel_class}
\alias{HyperModel_class}
\title{HyperModel}
\usage{
HyperModel_class(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
None
}
\description{
Defines a searchable space of Models and builds Models from this space.
}
