% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperResNet_HyperXception.R
\name{HyperXception}
\alias{HyperXception}
\title{HyperXception}
\usage{
HyperXception(
  include_top = TRUE,
  input_shape = NULL,
  input_tensor = NULL,
  classes = NULL
)
}
\arguments{
\item{include_top}{whether to include the fully-connected layer at the top of the network.}

\item{input_shape}{Optional shape list, e.g. `(256, 256, 3)`. One of `input_shape` or 
`input_tensor` must be specified.}

\item{input_tensor}{Optional Keras tensor (i.e. output of `layers.Input()`) to use as 
image input for the model. One of `input_shape` or `input_tensor` must be specified.}

\item{classes}{optional number of classes to classify images into, only to be specified 
if `include_top` is TRUE, and if no `weights` argument is specified. **kwargs: Additional 
keyword arguments that apply to all HyperModels. See `kerastuner.HyperModel`.}
}
\value{
a pre-trained Xception model
}
\description{
An Xception HyperModel.
}
