% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_kerastuner}
\alias{install_kerastuner}
\title{Install Keras Tuner}
\usage{
install_kerastuner(version = NULL, ..., restart_session = TRUE)
}
\arguments{
\item{version}{for specific version of Keras tuneR, e.g. "0.9.1"}

\item{...}{other arguments passed to [reticulate::py_install()].}

\item{restart_session}{Restart R session after installing (note this will only occur within RStudio).}
}
\value{
a python module kerastuner
}
\description{
This function is used to install the Keras Tuner python module
}
