% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{kerastuner}
\alias{kerastuner}
\title{Handle to the `kerastuner` module}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 5.
}
\usage{
kerastuner
}
\value{
Module(kerastuner)
}
\description{
Handle to the `kerastuner` module
}
\keyword{datasets}
