% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{load_model}
\alias{load_model}
\title{Load model}
\usage{
load_model(tuner, trial)
}
\arguments{
\item{tuner}{A tuner object}

\item{trial}{A `Trial` instance. For models that report intermediate results 
to the `Oracle`, generally `load_model` should load the best reported `step` 
by relying of `trial.best_step`}
}
\value{
None
}
\description{
Loads a Model from a given trial
}
