% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperclass.R
\name{Objective}
\alias{Objective}
\title{Objective}
\usage{
Objective(name = NULL, direction = NULL, ...)
}
\arguments{
\item{name}{name}

\item{direction}{direction}

\item{...}{Some additional arguments}
}
\description{
Objective(name, direction)
}
\details{

}
