% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tuner.R
\name{plot_tuner}
\alias{plot_tuner}
\title{Plot the tuner results with 'plotly'}
\usage{
plot_tuner(tuner, height = NULL, width = NULL, type = "plotly")
}
\arguments{
\item{tuner}{A tuner object}

\item{height}{height of the plot}

\item{width}{width of the plot}

\item{type}{Type parameter has 2 options: \cr 
* By default it uses `plotly` \cr 
* Second option is `echarts4r` \cr 
**Note** that `echarts4r` ignores width and height parameters}
}
\value{
a list which contains a dataframe of results and a plot
}
\description{
Plot the search space results
}
