% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{use_python}
\alias{tf}
\alias{PyClass}
\title{Objects exported from other packages}
\value{
a alias for reticulate::use_python

a alias for tensorflow::tf

a alias for reticulate::PyClass
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate]{PyClass}}, \code{\link[reticulate]{use_python}}}

  \item{tensorflow}{\code{\link[tensorflow]{tf}}}
}}

