% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{results_summary}
\alias{results_summary}
\title{Print a summary of the results}
\usage{
results_summary(tuner = NULL)
}
\arguments{
\item{tuner}{A tuner object}
}
\description{
Print a summary of the results
}
