% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_summary.R
\name{search_summary}
\alias{search_summary}
\title{Search summary}
\usage{
search_summary(tuner = NULL)
}
\arguments{
\item{tuner}{Requires a tuner object}
}
\description{
Print a summary of the search space
}
