% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_best_models}
\alias{get_best_models}
\title{The function for retrieving the top best models with hyperparameters}
\usage{
get_best_models(tuner = NULL, num_models = NULL)
}
\arguments{
\item{tuner}{A tuner object}

\item{num_models}{When search is over, one can retrieve the best model(s)}
}
\description{
The function for retrieving the top best models with hyperparameters
}
