% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperResNet_HyperXception.R
\name{Hyperband}
\alias{Hyperband}
\title{Tool for searching the best hyperparameters for computer vision.}
\usage{
Hyperband(
  hypermodel = NULL,
  objective = NULL,
  max_epochs = NULL,
  directory = NULL,
  project_name = NULL,
  ...
)
}
\arguments{
\item{hypermodel}{Define a model-building function. It takes an argument "hp" from which 
you can sample hyperparameters.}

\item{objective}{A loss metrics function for tracking the model performance e.g. "val_precision". The name of 
the objective to optimize (whether to minimize or maximize is automatically inferred for built-in metrics)}

\item{max_epochs}{to train the model. Note that in conjunction with initial_epoch, 
epochs is to be understood as "final epoch". The model is not trained for a number of iterations
 given by epochs, but merely until the epoch of index epochs is reached.}

\item{directory}{The dir where training logs are stored}

\item{project_name}{Detailed logs, checkpoints, etc, in the folder my_dir/helloworld, i.e. 
directory/project_name.}

\item{...}{Some additional arguments}
}
\description{
Tool for searching the best hyperparameters for computer vision.
}
