% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperclass.R
\name{HyperModel_class}
\alias{HyperModel_class}
\title{HyperModel}
\usage{
HyperModel_class()
}
\value{
None
}
\description{
Defines a searchable space of Models and builds Models from this space.
}
\details{
# Attributes: name: The name of this HyperModel. tunable: Whether the hyperparameters defined in this hypermodel 
should be added to search space. If `FALSE`, either the search space for these parameters must be defined in advance, 
or the default values will be used.
}
