% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperclass.R
\name{HyperModel_class}
\alias{HyperModel_class}
\title{You can use a HyperModel subclass instead of a model-building function
This makes it easy to share and reuse hypermodels.
A HyperModel subclass only needs to implement a build(self, hp) method.}
\usage{
HyperModel_class(object)
}
\description{
You can use a HyperModel subclass instead of a model-building function
This makes it easy to share and reuse hypermodels.
A HyperModel subclass only needs to implement a build(self, hp) method.
}
