% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperResNet_HyperXception.R
\name{HyperParameters}
\alias{HyperParameters}
\title{HyperParameters}
\usage{
HyperParameters(...)
}
\arguments{
\item{...}{Pass hyperparameter arguments to the tuner constructor}
}
\description{
The HyperParameters class serves as a hyerparameter container. A HyperParameters instance contains information about both the search space and the current values of each hyperparameter.
Hyperparameters can be defined inline with the model-building code that uses them. This saves you from having to write boilerplate code and helps to make the code more maintainable.
}
