% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperResNet_HyperXception.R
\name{HyperXception}
\alias{HyperXception}
\title{Xception is a convolutional neural network that is trained on more than a million images from the ImageNet 
database. The network is 71 layers deep and can classify images into 1000 object categories, such as keyboard,
mouse, pencil, and many animals.}
\usage{
HyperXception(input_shape = NULL, num_classes = NULL, ...)
}
\arguments{
\item{input_shape}{optional shape list, only to be specified if include_top is FALSE 
(otherwise the input shape has to be (224, 224, 3). It should have exactly 3 inputs channels, 
and width and height should be no smaller than 32. E.g. (200, 200, 3) would be one valid value.}

\item{num_classes}{optional number of classes to classify images into, only to be specified if include_top is True, 
and if no weights argument is specified.}

\item{...}{Some additional arguments}
}
\description{
These are ready-to-use hypermodels for computer vision.
}
\details{
Note: They come pre-compiled with loss="categorical_crossentropy" and metrics=["accuracy"].
}
