% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_kerastuner}
\alias{install_kerastuner}
\title{Install Keras Tuner}
\usage{
install_kerastuner(
  version = NULL,
  ...,
  bayesian = TRUE,
  restart_session = TRUE,
  from_git = FALSE
)
}
\arguments{
\item{version}{for specific version of Keras Tuner, e.g. "1.0.1"}

\item{...}{other arguments passed to [reticulate::py_install()].}

\item{bayesian}{install bayesian module}

\item{restart_session}{Restart R session after installing (note this will only occur within RStudio).}

\item{from_git}{install the recent GitHub version of Keras Tuner}
}
\value{
a python module kerastuner
}
\description{
This function is used to install the Keras Tuner python module
}
