% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_kerastuner}
\alias{install_kerastuner}
\title{Install Keras Tuner
This function is used to install the Keras Tuner python module on Mac OS and Linux (Windows will be added later)}
\usage{
install_kerastuner(python_path = NULL, upgrade = FALSE, version = NULL)
}
\arguments{
\item{python_path}{python_path specifies the python path for installation via terminal}

\item{upgrade}{is helpful if one wants to upgrade a version of Keras TuneR}

\item{version}{for specific version of Keras tuneR, e.g. "0.9.1"}
}
\description{
Install Keras Tuner
This function is used to install the Keras Tuner python module on Mac OS and Linux (Windows will be added later)
}
