% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperclass.R
\name{Objective}
\alias{Objective}
\title{Objective}
\usage{
Objective(name, direction, ...)
}
\arguments{
\item{name}{name}

\item{direction}{direction}

\item{...}{Some additional arguments}
}
\value{
None
}
\description{
Objective(name, direction) includes strings, 
the direction of the optimization (min or max) will be inferred.
}
