% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{Oracle}
\alias{Oracle}
\title{Oracle}
\usage{
Oracle(
  objective,
  max_trials = NULL,
  hyperparameters = NULL,
  allow_new_entries = TRUE,
  tune_new_entries = TRUE
)
}
\arguments{
\item{objective}{String. Name of model metric to minimize or maximize, e.g. "val_accuracy".}

\item{max_trials}{The maximum number of hyperparameter combinations to try.}

\item{hyperparameters}{HyperParameters class instance. Can be used to override (or register in 
advance) hyperparamters in the search space.}

\item{allow_new_entries}{Whether the hypermodel is allowed to request hyperparameter entries 
not listed in `hyperparameters`.}

\item{tune_new_entries}{Whether hyperparameter entries that are requested by the hypermodel 
but that were not specified in `hyperparameters` should be added to the search space, or not. 
If not, then the default value for these parameters will be used.}
}
\value{
None
}
\description{
Implements a hyperparameter optimization algorithm.
}
