% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tuner.R
\name{plot_tuner}
\alias{plot_tuner}
\title{Plot the tuner results with plotly}
\usage{
plot_tuner(tuner, height = NULL, width = NULL)
}
\arguments{
\item{tuner}{A tuner object}

\item{height}{height of the plot}

\item{width}{width of the plot}
}
\description{
Plot the search space results
}
