% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_summary.R
\name{results_summary}
\alias{results_summary}
\title{Results summary}
\usage{
results_summary(tuner = NULL, num_trials = NULL)
}
\arguments{
\item{tuner}{Requires a tuner object}

\item{num_trials}{Shows the top best models}
}
\value{
the list of results summary of the tuner object
}
\description{
Print a summary of the search results (best models)
}
