% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperclass.R
\name{Tuner_class}
\alias{Tuner_class}
\title{Tuner class for Keras models.
May be subclassed to create new tuners.}
\usage{
Tuner_class()
}
\value{
None
}
\description{
Tuner class for Keras models.
May be subclassed to create new tuners.
}
